% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_full_posterior}
\alias{get_full_posterior}
\title{Approximation Posterior Expectations from Mr.ASH Fit}
\usage{
get_full_posterior(X, y, w, beta, pi, sigma2, sa2)
}
\arguments{
\item{fit}{A Mr.ASH fit obtained, for example, by running
\code{\link{mr_ash}}.}
}
\value{
A list object with the following elements:

\item{phi}{A p x K matrix containing the posterior assignment
  probabilities, where p is the number of predictors, and K is the
  number of mixture components. (Each row of \code{phi} should sum to
  1.)}

\item{m}{A p x K matrix containing the posterior means conditional
  on assignment to each mixture component.}

\item{s2}{A p x K matrix containing the posterior variances
  conditional on assignment to each mixture component.}

\item{lfsr}{A vector of length p containing the local false
  discovery rate for each variable}
}
\description{
Recover the parameters specifying the variational
  approximation to the posterior distribution of the regression
  coefficients. To streamline the model fitting implementation, and
  to reduce memory requirements, \code{\link{mr_ash}} does not store
  all the parameters needed to specify the approximate posterior.
}
\examples{
## generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

# Fit mr.ash model.
fit <- mr_ash(X, y)

}
