% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{simulate_regression_data}
\alias{simulate_regression_data}
\title{Simulate Data from Multiple Regression Model}
\usage{
simulate_regression_data(
  n,
  p,
  s,
  sigma = 1,
  pve = 0.5,
  center_X = TRUE,
  standardize_X = FALSE,
  intercept = 0,
  ncov = 0
)
}
\arguments{
\item{n}{number of samples. Must be at least 2.}

\item{p}{number of regression variables. Must be at least 2.}

\item{s}{number of non-zero regression coefficients.}

\item{sigma}{standard deviation of the residual.}

\item{pve}{proportion of variance in Y explained by X.}

\item{center_X}{If \code{center_X = TRUE}, the columns of X are
centered before simulating y so that each column as a mean of
zero.}

\item{standardize_X}{If \code{standardize_X = TRUE}, the columns of
X are standardized before simulating y so that each column as a
mean of zero and a standard deviation of 1.}

\item{intercept}{intercept of regression equation. \eqn{\beta_0} below.}

\item{ncov}{(optional) number of covariates in Z. Defaults to 0}
}
\value{
A list with components
\itemize{
  \item y - \code{n} vector containing regression response.
  \item X - \code{n} by \code{p} matrix of regression variables X.
  \item Z - \code{n} by \code{ncov} matrix of covariates Z.
  \item u - \code{ncov} vector containing covariate coefficients u.
  \item beta - \code{p} vector contained regression coefficients \eqn{\beta}.
}
}
\description{
Simulate regression data from a multiple regression
  model.
}
\details{
Data are generated from the model \deqn{y | X, \beta,
\sigma^2, \beta_0 \sim N(\beta_0 + X \beta + Zu, \sigma I_n),} in
which the non-zero regression coefficients are drawn from the
standard normal, and then scaled to achieve the desired proportion
of variance in y explained by X (see input argument \code{pve}).
}
\examples{

# simulate 100 x 100 regression data with 75\% sparsity
simulate_regression_data(100, 100, 25)

}
